requirements.txt
opencv-python
numpy
numba
torch
torchvision
Pillow
pytorch-lightning
hydra-core
python-dotenv
